requirements.txt
arabic-reshaper==2.1.0
asgiref==3.2.10
astroid==2.4.2
colorama==0.4.4
cycler==0.10.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.1.2
django-cors-headers==3.5.0
django-filter==2.4.0
django-import-export==2.4.0
django-jalali==4.0.0
djangorestframework==3.12.1
et-xmlfile==1.0.1
future==0.18.2
isort==5.6.4
jdatetime==3.6.2
jdcal==1.4.1
kiwisolver==1.3.1
lazy-object-proxy==1.4.3
MarkupPy==1.14
matplotlib==3.3.3
mccabe==0.6.1
numpy==1.19.3
odfpy==1.4.1
openpyxl==3.0.5
pandas==1.1.4
Pillow==8.0.1
psycopg2==2.8.6
pycode-similar==1.4
pylint==2.6.0
pyparsing==2.4.7
python-bidi==0.4.2
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
sqlparse==0.4.1
tablib==3.0.0
termcolor==1.1.0
toml==0.10.1
wrapt==1.12.1
xlrd==1.2.0
xlwt==1.3.0
